Working on minimizing allocations of Struct
mwe_structs*/ : experiments to better understand when allocations occur
when using structs  (mutable vs immutable). So far, I find that
non-parametrized structures have less allocations. I do not know the
reason. Maybe because in that case, allocation in on the heap, and the
individual structs can be accessed via pointer. If they are allocated on
the stack, the might be accessed faster, but when there is a need to
access the data, an allocation must be made (but I do not know why).